requirements.txt
azure.eventhub
azure.identity
aiofiles
datetime